Invert flake inputs
This prioritizes Flaky’s inputs, and triggers an upgrade to Nixpkgs 24.05.

There are a number of fixes related to that upgrade. Among them is removing
Emacs’ direnv package. I was already using envrc instead, but had the old direnv
config laying around. The patches for direnv were out of date, so finally
chopped it out.