Make it easier to define systems downstream
Adds functions that set up all the `specialArgs`, which means downstream
configurations only need to include ones that they explicitly depend on and
don’t need to duplicate so many flake inputs.

This also removes `stateVersion` from the configurations, as that should be set
by the final configuration downstream, and only bumped once that configuration
is confident it has addressed any possible issues.