Move Skitch to mas (darwin) from nixcasks (home)
Skitch is likely disappearing. Evernote has announced that it’s discontinuing
it, and [Homebrew has correspondingly removed
it](https://github.com/orgs/Homebrew/discussions/6243). However, at least for
now, it’s still available on the App Store, so make it a system-wide package
instead of a user-specific package, and keep on.